workflows/pyproject.toml
[project]
name = "foundry-hosted-langchain-workflow-demos"
version = "0.1.0"
description = "LangGraph workflow demos for Microsoft Foundry"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-agentserver-responses==1.0.0b5",
    "azure-ai-agentserver-core==2.0.0b3",
    "azure-identity>=1.25.3",
    "azure-monitor-opentelemetry>=1.8.7",
    "langchain-azure-ai[opentelemetry]>=1.2.3",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.9",
    "python-dotenv>=1.2.2",
    "vendor-langchain-azure-ai-runtime",
]

[tool.uv]
prerelease = "allow"

[tool.uv.sources]
vendor-langchain-azure-ai-runtime = { path = "../vendor" }

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = ["vendor"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
